requirements.txt
confluent-kafka==2.6.0
python-dotenv==1.0.1
requests==2.32.3
pydantic==2.9.2
langchain==0.3.2
langchain-openai==0.2.2
langchain-community==0.3.1
tiktoken==0.7.0
